+++ apputils/eval.go
@@ -1,9 +1,10 @@
-package utils
+package apputils
 
 import (
+	"github.com/pkg/errors"
+	"go.mongodb.org/mongo-driver/bson"
 	"reflect"
 	"regexp"
-
 )
 
 // Eval is a typed representation of logic that can be evaluated.
